pyproject.toml
[project]
name = "daily-briefing"
version = "0.1.0"
description = "AWS Lambda function that generates personalized daily briefings using Claude Opus 4.5"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic==0.43.1",
    "boto3==1.35.97",
]

[project.optional-dependencies]
dev = [
    "aws-cdk-lib==2.175.2",
    "constructs>=10.0.0,<11.0.0",
    "pytest==8.3.5",
    "pytest-cov==6.0.0",
    "pytest-mock==3.14.0",
    "moto[all]==5.0.31",
    "black==25.1.0",
    "flake8==7.1.1",
    "mypy==1.14.1",
    "pylint==3.3.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
